split generic elastic Client logic from our elastic endpoints
