+++ web-api/src/storage/elastic.rs
@@ -12,24 +12,18 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+mod client;
+
 use std::collections::HashMap;
 
+pub(crate) use client::{Client, ClientBuilder, Config};
 use itertools::Itertools;
-use reqwest::{
-    header::{HeaderMap, HeaderValue, CONTENT_TYPE},
-    Body,
-    StatusCode,
-    Url,
-};
-use secrecy::{ExposeSecret, Secret};
-use serde::{de::DeserializeOwned, Deserialize, Deserializer, Serialize};
+use serde::{Deserialize, Deserializer, Serialize};
 use serde_json::{json, Value};
-use tracing::error;
 use xayn_ai_bert::NormalizedEmbedding;
 
+use self::client::BulkInstruction;
 use crate::{
-    app::SetupError,
-    error::common::InternalError,
     models::{
         self,
         DocumentId,
@@ -39,117 +33,8 @@ use crate::{
         DocumentTag,
     },
     storage::{KnnSearchParams, Warning},
-    utils::{serialize_redacted, serialize_to_ndjson},
     Error,
 };
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-#[serde(default)]
-pub(crate) struct Config {
-    url: String,
-    user: String,
-    #[serde(serialize_with = "serialize_redacted")]
-    password: Secret<String>,
-    index_name: String,
-}
-
-impl Default for Config {
-    fn default() -> Self {
-        Self {
-            url: "http://localhost:9200".into(),
-            user: "elastic".into(),
-            password: String::from("changeme").into(),
-            index_name: "test_index".into(),
-        }
-    }
-}
-
-#[derive(Clone)]
-pub(crate) struct Client {
-    config: Config,
-    url_to_index: Url,
-    client: reqwest::Client,
-}
-
-impl Client {
-    pub(crate) fn new(config: &Config) -> Result<Self, SetupError> {
-        let mut url_to_index = config.url.parse::<Url>()?;
-        url_to_index
-            .path_segments_mut()
-            .map_err(|()| anyhow::anyhow!("non segmentable url in config"))?
-            .push(&config.index_name);
-
-        Ok(Self {
-            config: config.clone(),
-            url_to_index,
-            client: reqwest::Client::new(),
-        })
-    }
-}
-
-#[derive(Debug, Serialize)]
-#[serde(rename_all = "lowercase")]
-enum BulkInstruction<'a> {
-    Index {
-        #[serde(rename = "_id")]
-        id: &'a DocumentId,
-    },
-    Delete {
-        #[serde(rename = "_id")]
-        id: &'a DocumentId,
-    },
-}
-
-#[derive(Debug, Deserialize)]
-struct BulkItemResponse {
-    #[serde(rename = "_id")]
-    id: DocumentId,
-    status: u16,
-    #[serde(default)]
-    error: Value,
-}
-
-impl BulkItemResponse {
-    fn is_success_status(&self, allow_not_found: bool) -> bool {
-        StatusCode::from_u16(self.status)
-            .map(|status| {
-                (status == StatusCode::NOT_FOUND && allow_not_found) || status.is_success()
-            })
-            .unwrap_or_default()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-struct BulkResponse {
-    errors: bool,
-    items: Vec<HashMap<String, BulkItemResponse>>,
-}
-
-impl BulkResponse {
-    fn failed_documents(self, operation: &'static str, allow_not_found: bool) -> Vec<DocumentId> {
-        self.errors.then(|| {
-            self
-                .items
-                .into_iter()
-                .filter_map(|mut response| {
-                    if let Some(response) = response.remove(operation) {
-                        if !response.is_success_status(allow_not_found) {
-                            error!(
-                                document_id=%response.id,
-                                error=%response.error,
-                                "Elastic failed to {operation} document.",
-                            );
-                            return Some(response.id);
-                        }
-                    } else {
-                        error!("Bulk {operation} request contains non {operation} responses: {response:?}");
-                    }
-                    None
-                })
-                .collect()
-        }).unwrap_or_default()
-    }
-}
 
 #[derive(Debug, Deserialize)]
 /// Deserializes from any map/struct dropping all fields.
@@ -159,102 +44,6 @@ impl BulkResponse {
 struct IgnoredResponse {/* Note: The {} is needed for it to work correctly. */}
 
 impl Client {
-    fn create_resource_path<'a>(
-        &self,
-        segments: impl IntoIterator<Item = &'a str>,
-        query_parts: impl IntoIterator<Item = (&'a str, Option<&'a str>)>,
-    ) -> Url {
-        let mut url = self.url_to_index.clone();
-        // UNWRAP_SAFE: In the constructor we already made sure it's a segmentable url.
-        url.path_segments_mut().unwrap().extend(segments);
-        let mut query_mut = url.query_pairs_mut();
-        for (key, value) in query_parts {
-            if let Some(value) = value {
-                query_mut.append_pair(key, value);
-            } else {
-                query_mut.append_key_only(key);
-            }
-        }
-        drop(query_mut);
-
-        url
-    }
-
-    async fn bulk_request(
-        &self,
-        requests: impl IntoIterator<Item = Result<impl Serialize, Error>>,
-    ) -> Result<BulkResponse, Error> {
-        let url = self.create_resource_path(["_bulk"], [("refresh", None)]);
-
-        let mut headers = HeaderMap::new();
-        headers.insert(
-            CONTENT_TYPE,
-            HeaderValue::from_static("application/x-ndjson"),
-        );
-
-        let body = serialize_to_ndjson(requests)?;
-
-        self.query_with_bytes::<_, BulkResponse>(url, Some((headers, body)))
-            .await?
-            .ok_or_else(|| InternalError::from_message("_bulk endpoint not found").into())
-    }
-
-    async fn query_with_bytes<B, T>(
-        &self,
-        url: Url,
-        post_data: Option<(HeaderMap<HeaderValue>, B)>,
-    ) -> Result<Option<T>, Error>
-    where
-        B: Into<Body>,
-        T: DeserializeOwned,
-    {
-        let request_builder = if let Some((headers, body)) = post_data {
-            self.client.post(url).headers(headers).body(body)
-        } else {
-            self.client.get(url)
-        };
-
-        let response = request_builder
-            .basic_auth(
-                &self.config.user,
-                Some(self.config.password.expose_secret()),
-            )
-            .send()
-            .await?;
-
-        let status = response.status();
-        if status == StatusCode::NOT_FOUND {
-            Ok(None)
-        } else if !status.is_success() {
-            let url = response.url().clone();
-            let body = response.bytes().await?;
-            let err_msg = String::from_utf8_lossy(&body);
-            Err(InternalError::from_message(format!(
-                "Elastic Search failed, status={status}, url={url}, \nbody={err_msg}"
-            ))
-            .into())
-        } else {
-            Ok(Some(response.json().await?))
-        }
-    }
-
-    async fn query_with_json<B, T>(&self, url: Url, body: Option<B>) -> Result<Option<T>, Error>
-    where
-        B: Serialize,
-        T: DeserializeOwned,
-    {
-        let post_data = body
-            .map(|json| -> Result<_, Error> {
-                let mut headers = HeaderMap::new();
-                headers.insert(CONTENT_TYPE, HeaderValue::from_static("application/json"));
-                let body = serde_json::to_vec(&json)?;
-                Ok((headers, body))
-            })
-            .transpose()?;
-
-        self.query_with_bytes(url, post_data).await
-    }
-
     #[allow(clippy::too_many_lines)]
     pub(super) async fn get_by_embedding<'a>(
         &self,
